@@ -25,8 +25,8 @@
         "dist"
     ],
     "peerDependencies": {
-        "@sveltejs/kit": "1.x || 2.x",
-        "svelte": "3.x || 4.x"
+        "@sveltejs/kit": ">= 2.5.27",
+        "svelte": ">= 5"
     },
     "peerDependenciesMeta": {
         "@sveltejs/kit": {
@@ -35,23 +35,23 @@
     },
     "devDependencies": {
         "@sveltejs/adapter-vercel": "5.1.0",
-        "@sveltejs/kit": "2.5.0",
+        "@sveltejs/kit": "2.5.27",
         "@sveltejs/package": "2.2.6",
-        "@sveltejs/vite-plugin-svelte": "3.0.2",
+        "@sveltejs/vite-plugin-svelte": "4.0.0",
         "autoprefixer": "10.4.17",
         "cssnano": "6.0.3",
         "postcss": "8.4.35",
         "postcss-load-config": "5.0.3",
         "prettier": "3.2.5",
-        "prettier-plugin-svelte": "3.2.1",
+        "prettier-plugin-svelte": "3.2.6",
         "prettier-plugin-tailwindcss": "0.5.11",
         "publint": "0.2.7",
-        "svelte": "4.2.11",
-        "svelte-check": "3.6.4",
+        "svelte": "5.0.0",
+        "svelte-check": "4.0.0",
         "tailwindcss": "3.4.1",
         "tslib": "2.6.2",
-        "typescript": "5.3.3",
-        "vite": "5.1.3"
+        "typescript": "5.6.3",
+        "vite": "5.4.4"
     },
     "types": "./dist/index.d.ts",
     "svelte": "./dist/index.js",
